pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "uranium-model"
version = "0.1.0"
description = "Uranium fuel-cycle supply-demand simulator with explainable price layer"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "SQLAlchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "statsmodels>=0.14.0",
    "matplotlib>=3.7.0",
    "plotly>=5.20.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
packages = [
    "uranium_model",
    "uranium_model.connections",
    "uranium_model.data",
    "uranium_model.core",
    "uranium_model.features",
    "uranium_model.models",
    "uranium_model.config",
    "uranium_model.analysis",
]

[tool.setuptools.package-data]
uranium_model = ["**/*.yaml", "**/*.yml", "**/*.csv"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
